Replace slong by ulong for safety in ulong_extras
For example, shifting WORD(1) << 63 is undefined.

+++ ulong_extras/powmod2_preinv.c
@@ -31,7 +31,7 @@ n_powmod2_preinv(ulong a, slong exp, ulong n, ulong ninv)
             flint_throw(FLINT_IMPINV, "Cannot invert modulo %wd*%wd\n", g,
                         n / g);
 
-        exp = -exp;
+        exp = -((ulong) exp);
     }
 
     count_leading_zeros(norm, n);
