+++ ulong_extras/test/t-powmod2_preinv.c
@@ -47,7 +47,7 @@ main(void)
         flint_mpz_set_ui(d_m, d);
         if (exp < 0)
         {
-            flint_mpz_powm_ui(r2_m, a_m, -exp, d_m);
+            flint_mpz_powm_ui(r2_m, a_m, -((ulong) exp), d_m);
             mpz_invert(r2_m, r2_m, d_m);
         }
         else
